pyproject.toml
[project]
name = "RedBot"
version = "0.1.0"
description = "一个使用 Graia Ariadne 搭建的 QQ 机器人"
authors = [
    { name = "Redlnn", email = "w731347477@gmail.com" },
]
dependencies = [
    "aio-mc-rcon",
    "aiofile",
    "aiohttp[speedups]",
    "asyncmy",
    "aiosqlite",
    "graia-ariadne[full]",
    "graiax-fastapi",
    "graiax-playwright",
    "graiax-shortcut",
    "graiax-text2img-playwright",
    "jieba",
    "jinja2",
    "lxml",
    "numpy",
    "orjson",
    "passlib[bcrypt]",
    "pillow",
    "psutil",
    "python-jose[cryptography]",
    "python-multipart",
    "qrcode",
    "sqlalchemy[asyncio]",
    "uvicorn[standard]",
    "wordcloud",
    "kayaku",
]
requires-python = ">=3.10,<4.0"
license = { text = "AGPL-3.0-only" }
readme = "README.md"

[project.urls]
repository = "https://github.com/Redlnn/redbot"

[tool.pdm]
[tool.pdm.dev-dependencies]
dev = [
    "isort",
    "black",
    "flake8",
    "pip",
    "wheel",
    "setuptools",
]

[build-system]
requires = ["pdm-pep517>=1.0.0"]
build-backend = "pdm.pep517.api"

[tool.black]
line-length = 120
target-version = ["py310"]
skip-string-normalization = true
safe = true

[tool.isort]
profile = "black"
